// src/models/TopPlaysfilter.ts
export class TopPlaysfilter {
    gamemode: string = "";
    search: string = "";
    offset: number = 0;
    mods: Array<string> = [];
    rank: Array<string> = [];
    combo: Array<string> = [];
    acc: Array<string> = [];
    sort: string = "position";
    reverse: boolean = true;
    discordid: string = "";
    username: string = "";
}